refactor: dto 정리
- dto 디렉토리 변경
