--- src/main/java/com/example/naejango/domain/chat/dto/request/CreatePrivateChatRequestDto.java
@@ -1,14 +1,0 @@
-package com.example.naejango.domain.chat.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class CreatePrivateChatRequestDto {
-    private Long receiverId;
-}
